fix(tsdf): close the whole-branch review gate on hash strategy
Nine review findings, one wave.

- tsdf_folder_eval: restore scene-size auto-tiling. --tsdf now defaults to
  tile above 512 voxels/axis and flat below, decided at runtime once the
  frames are read, instead of a declared "flat" that silently truncated
  oversized scans to a 512^3 sliver (scan_out --voxel 0.01 extracted ZERO
  points and still printed a healthy stats table). An explicit --tsdf flat
  on such a scene is still honoured but now warns on stderr, and --single's
  error text no longer points at the removed auto-tile path.
- BucketedStoresTheSameEntriesAsLinear now compares EXTRACTION, not just
  occupiedEntryCount. findSlot is reached only from the extract kernel, so
  the old assertions were byte-identical under any lookup defect; the test
  now diffs extracted point counts and sorted positions under both hashes.
  Verified by injecting a findSlot defect: the new assertion fails (0 vs
  169 points) while all three old ones still pass.
- Name bucketed's probe budget HASH_MAX_BUCKET_PROBE and set it to
  MAX_PROBE / HASH_BUCKET_SIZE, so both strategies give up after the same
  128 SLOTS instead of bucketed sweeping 4096. Both loops state the budget
  in slots. This is a measured behaviour change: at --tile-hash 8192
  bucketed now reports 739 drops where the 32x budget hid them (table
  memory unchanged), which is exactly the false parity the review flagged.
  Loop variables renamed off abbreviations (bucketStep/slotInBucket/...).
- Document what a 0.8 load factor costs beside loadFactorLimit: maybeGrow
  decides from a host FilledCount() reflecting only completed submissions,
  so growth is sized to stale occupancy and a frame inserting more than
  (1 - limit) * capacity new entries drops the surplus irrecoverably --
  half the table of headroom at 0.5, a fifth at 0.8.
- SubmapStrategy: name the detail truncation constant instead of a bare 3.0f.
- Sweep stale advanced_tsdf_* kernel filenames left by the move (headers,
  kernels, Spatial/README table, tsdf_benchmark, and two dead doc links).
- Drop the unread HASH_NAME from both fragments; HashStrategy::name is the
  single name for the concept.
- Sum InsertFailureCount/GrowCount into uint64_t across tiles and levels,
  and widen VolumeStats::growCount to match its insertFailureCount sibling.

Full suite: 287 passed / 1 skipped / 0 failed, unchanged from baseline.

+++ src/TSDF/Backends/SubmapAdvancedTSDF.h
@@ -157,11 +157,13 @@ namespace TSDF {
         uint32_t BaseTileCount() const { return m_base.TileCount(); }
         uint32_t DetailTileCount() const { return m_detail.TileCount(); }
 
-        uint32_t InsertFailureCount() const {
+        // 64-bit like the tiled levels they sum: both levels are full tile hierarchies, so the
+        // totals are the two largest tile counts in the system added together.
+        uint64_t InsertFailureCount() const {
             return m_base.InsertFailureCount() + m_detail.InsertFailureCount();
         }
 
-        uint32_t GrowCount() const { return m_base.GrowCount() + m_detail.GrowCount(); }
+        uint64_t GrowCount() const { return m_base.GrowCount() + m_detail.GrowCount(); }
 
         void DownloadEntries(std::vector<AdvancedEntry> &out) const {
             if (m_dense.empty()) {
